requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
langchain
langchain-community
langchain-core
langchain-chroma
langchain-huggingface
langchain-groq
langchain-google-genai
unstructured
unstructured[all-docs]
huggingface-hub
chromadb
sentence-transformers
